pyproject.toml
[project]
name = "ai-for-data-quality-control"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "argparse>=1.4.0",
    "fastparquet>=2024.11.0",
    "jupyter>=1.1.1",
    "logger>=1.4",
    "logging>=0.4.9.6",
    "matplotlib>=3.10.7",
    "mlflow>=3.7.0",
    "numpy>=2.3.4",
    "nvidia-ml-py>=13.590.44",
    "pandas>=2.3.3",
    "pathlib>=1.0.1",
    "pip>=25.3",
    "plotly>=6.5.0",
    "psutil>=7.1.3",
    "pyarrow>=22.0.0",
    "pynvml>=13.0.1",
    "pytest>=8.4.2",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "streamlit>=1.52.1",
    "torch>=2.9.1",
    "torchmetrics>=1.8.2",
    "torchvision>=0.24.1",
    "tqdm>=4.67.1",
]

[dependency-groups]
dev = [
    "mkdocs>=1.6.1",
    "mkdocs-material>=9.6.23",
    "mkdocstrings[python]>=0.30.1",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["models*"]
exclude = ["mlruns*", "runs*", "sandbox*", "params*"]
